requirements.txt
django>=1.11
django-debug-toolbar=1.11
django-pivot==1.8.0
django-sql-utils==0.1.1
Pygments==2.3.1
sqlparse==0.3.0